requirements.txt
requests
pyyaml
jq
beautifulsoup4
jinja2